PMM-7922 rename MultiSelect
