--- public/app/percona/shared/components/Form/MultiSelect/index.ts
@@ -1,1 +1,0 @@
-export * from './MultiSelect';
