+++ public/app/percona/shared/components/Form/MultiSelectField/index.ts
@@ -1,0 +1,1 @@
+export * from './MultiSelectField';
